confirmação de deleção, erro carrega limpa lista, remoção logs

+++ src/app/components/atividadeludica/atividadeludica-listar/atividadeludica-listar.component.ts
@@ -5,14 +5,14 @@ import { CrudService } from 'src/app/shared/crud-service/crud-service';
 import { Router } from '@angular/router';
 import { LabelValue } from 'src/app/shared/labelvalue/labelvalue';
 import { Situacoes } from 'src/app/shared/situacoes/situacoes';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { ToastService } from 'src/app/shared/toast-service/toast.service';
 
 @Component({
   selector: 'app-atividadeludica',
   templateUrl: './atividadeludica-listar.component.html',
   styleUrls: ['./atividadeludica-listar.component.css'],
-  providers: [{ provide: CrudService, useExisting: AtividadeLudicaService }]
+  providers: [{ provide: CrudService, useExisting: AtividadeLudicaService }, ConfirmationService]
 })
 export class AtividadeludicaListarComponent {
 
@@ -33,7 +33,8 @@ export class AtividadeludicaListarComponent {
   constructor(
     private service: CrudService<AtividadeLudica>,
     private router: Router,
-    private toastService: ToastService
+    private toastService: ToastService,
+    private confirmationService: ConfirmationService
   ) { }
 
   novo() {
@@ -44,9 +45,23 @@ export class AtividadeludicaListarComponent {
   }
 
   delete(id: any) {
-    this.service.delete(id).subscribe((res) => {
-      this.carregarLista(0, this.rows);
-    })
+    this.confirmationService.confirm({
+      message: 'Tem certeza que deseja deletar?',
+      header: 'Confirmação',
+      icon: 'pi pi-info-circle',
+      acceptButtonStyleClass: "p-button-danger p-button-text",
+      rejectButtonStyleClass: "p-button-text p-button-text",
+      acceptIcon: "none",
+      rejectIcon: "none",
+      accept: () => {
+        this.service.delete(id).subscribe((res) => {
+          this.carregarLista(0, this.rows);
+        })
+      },
+      reject: () => {
+      }
+    });
+
   }
 
   edit(id: number) {
@@ -63,6 +78,8 @@ export class AtividadeludicaListarComponent {
       },
       error: (error) => {
         this.carregando = false;
+        this.total = 0;
+        this.items = [];
         this.toastService.toastBase('warn', 'Não foi possível carregar!', error);
       }
     })
