+++ src/app/components/entradasaida/entradasaida-listar/entradasaida-listar.component.ts
@@ -4,14 +4,14 @@ import { CrudService } from 'src/app/shared/crud-service/crud-service';
 import { EntradaSaida } from '../modelo/entradasaida';
 import { Router } from '@angular/router';
 import { ResidenteService } from '../../residente/service/residente.service';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { ToastService } from 'src/app/shared/toast-service/toast.service';
 
 @Component({
   selector: 'app-entradasaida-listar',
   templateUrl: './entradasaida-listar.component.html',
   styleUrls: ['./entradasaida-listar.component.css'],
-  providers: [{ provide: CrudService, useExisting: EntradasaidaService }]
+  providers: [{ provide: CrudService, useExisting: EntradasaidaService }, ConfirmationService]
 
 })
 export class EntradasaidaListarComponent {
@@ -30,7 +30,8 @@ export class EntradasaidaListarComponent {
   constructor(
     private service: CrudService<EntradaSaida>,
     private router: Router,
-    private toastService: ToastService
+    private toastService: ToastService,
+    private confirmationService: ConfirmationService
   ) {
   }
 
@@ -42,9 +43,22 @@ export class EntradasaidaListarComponent {
   }
 
   delete(id: any) {
-    this.service.delete(id).subscribe((res) => {
-      this.carregarLista(0, this.rows);
-    })
+    this.confirmationService.confirm({
+      message: 'Tem certeza que deseja deletar?',
+      header: 'Confirmação',
+      icon: 'pi pi-info-circle',
+      acceptButtonStyleClass: "p-button-danger p-button-text",
+      rejectButtonStyleClass: "p-button-text p-button-text",
+      acceptIcon: "none",
+      rejectIcon: "none",
+      accept: () => {
+        this.service.delete(id).subscribe((res) => {
+          this.carregarLista(0, this.rows);
+        })
+      },
+      reject: () => {
+      }
+    });
   }
 
   edit(id: number) {
@@ -61,6 +75,8 @@ export class EntradasaidaListarComponent {
       },
       error: (error) => {
         this.carregando = false;
+        this.total = 0;
+        this.items = [];
         this.toastService.toastBase('warn', 'Não foi possível carregar!', error);
       }
     })
